// src/app/core/models/share/query.model.ts
export class Query {
  clubKey?: string;
  type?: string;
  startDate?: string;
  endDate?: string;
  limit?: number;
  paymentType?: string;
  order?: string;
  isFrozen?: boolean;
  clubId?: number;
  paymentPlanId?: number;
  ptPackageId?: number;
  status?: string;
  page?: number;
  queryString?: string;
  accountId?: string;
  paymentMethod?: string;
  isExport?: boolean;
  ptId?: number;
  ptGroupId?: number;
  groupId?: number;
  leaderId?: number;
  orderKey?: string;
  orderType?: string;
  chargeFailCount?: number;
  isActive?: number;
  startCreatedDate?: string;
  endCreatedDate?: string;
  expiredDay?: number;
  numRemainingDay?: number;
  numUnuseDay?: number;
  endPoint?: number;
  categoryId?: number;
  warehouseId?: number;
  classId?: number;
  roleId?: number;
  supplierId?: number;
  startCreatedAt?: string;
  endCreatedAt?: string;
  startApprovedAt?: string;
  endApprovedAt?: string;
  startConfirmedAt?: string;
  endConfirmedAt?: string;
  typeTransaction?: string;
  isTotal?: boolean;
  isAll?: boolean;
  transactionType?: string;
}